Move grafana-specific scenes under dedicated folder
