+++ public/app/features/alerting/unified/insights/grafana/FiringAlertsPercentage.tsx
