+++ public/app/features/alerting/unified/insights/grafana/FiringAlertsRate.tsx
