+++ public/app/features/alerting/unified/insights/grafana/MostFiredRulesTable.tsx
